requirements.txt
# Add your dependencies here

## MedAgent
openai==0.27.4
torch==2.0.0
nltk==3.8.1
numpy==1.23.3
rouge-score==0.1.2
wrapt==1.15.0
wrapt-timeout-decorator==1.4.0
jsonlines==4.0.0
tqdm==4.65.0
python-dotenv==1.0.1


## Self-BioRAG